Refac: inclusions/requires
I moved Swarm.Text into a separate file as it is a useful
class that has to evolve on its own. Still, it is not
supposed to be used by every single project out there, so
it goes to a separate file.
The inclusion system now handles node.js require()s and
old plain <script> inclusion. Everything else is better
done by grunt/browserify, not guessed at run time.